requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
nltk
Pillow
streamlit-webrtc
tensorflow-cpu